refactor PostProcessAdmin ::reduce() and ::create_reduction_script() to child classes of BaseProcessor
